# api/openapi.yaml
openapi: 3.1.0
info:
  title: AI-CRM v3 Foundation API
  version: 0.4.0
  description: OneID, employee access, durable WeCom callback and customer directory, External Effects, and Media contracts.
servers:
  - url: https://id-dev.youcangogogo.com
paths:
...
  /api/admin/orders:
    get:
      operationId: listLegacyOrders
      tags: [Transactions]
      security: [{adminSession: []}]
      parameters:
        - {$ref: "#/components/parameters/Limit"}
        - {name: cursor, in: query, schema: {type: string, maxLength: 512}}
        - {name: provider, in: query, schema: {type: string, enum: [wechat, wechat_pay, wechat_shop, alipay]}}
        - {name: status, in: query, schema: {type: string, enum: [pending_payment, paid, partially_refunded, refunded, cancelled, payment_failed, closed]}}
        - {name: order_ref, in: query, schema: {type: string, maxLength: 200}}
        - {name: customer_id, in: query, schema: {$ref: "#/components/schemas/PositiveID"}}
        - {name: product, in: query, schema: {type: string, maxLength: 200}}
        - {name: created_from, in: query, schema: {type: string, format: date}}
        - {name: created_to, in: query, schema: {type: string, format: date}}
      responses: {"200": {description: Canonical order page, content: {application/json: {schema: {$ref: "#/components/schemas/TransactionOrderPage"}}}}, "400": {$ref: "#/components/responses/Error"}, "401": {$ref: "#/components/responses/Error"}, "403": {$ref: "#/components/responses/Error"}, "409": {$ref: "#/components/responses/Error"}, "503": {$ref: "#/components/responses/Unavailable"}}
...
  /api/admin/refunds:
    get:
      operationId: listLegacyRefunds
      tags: [Transactions]
      security: [{adminSession: []}]
      responses: {"200": {description: Refund page, content: {application/json: {schema: {type: object, additionalProperties: true}}}}, "401": {$ref: "#/components/responses/Error"}, "403": {$ref: "#/components/responses/Error"}}
    post:
      operationId: createWechatShopRefundIntent
      tags: [Transactions]
      security: [{adminSession: [], csrfHeader: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [provider, order_no, product_id, sku_id, refund_count, refund_amount_total, reason_code, reason, transaction_id_confirmation, checked]
              properties:
                provider: {type: string, enum: [wechat_shop]}
                order_no: {type: string, minLength: 1, maxLength: 200}
                product_id: {type: string, minLength: 1, maxLength: 200}
                sku_id: {type: string, minLength: 1, maxLength: 200}
                refund_count: {type: integer, minimum: 1, maximum: 1000000}
                refund_amount_total: {type: integer, minimum: 1}
                reason_code: {type: string, enum: ["10000000", "10000001", "10000002", "10000006", "10000007", "10000008", "10000014", "10000015", "10000017", "10000021"]}
                reason: {type: string, minLength: 1, maxLength: 500}
                transaction_id_confirmation: {type: string, minLength: 1, maxLength: 200}
                checked: {type: boolean}
                operator: {type: string, maxLength: 200}
      responses: {"202": {description: WeChat Shop refund effect accepted atomically}, "400": {$ref: "#/components/responses/Error"}, "401": {$ref: "#/components/responses/Error"}, "403": {$ref: "#/components/responses/Error"}, "404": {$ref: "#/components/responses/Error"}, "409": {$ref: "#/components/responses/Error"}, "503": {$ref: "#/components/responses/Unavailable"}}
  /api/admin/exports/preview:
    post:
      operationId: previewLegacyOrderExport
      tags: [Transactions]
      security: [{adminSession: [], csrfHeader: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/TransactionExportRequest"}}}}
      responses: {"200": {description: Bounded export preview, content: {application/json: {schema: {type: object, additionalProperties: true}}}}, "400": {$ref: "#/components/responses/Error"}, "401": {$ref: "#/components/responses/Error"}, "403": {$ref: "#/components/responses/Error"}}
  /api/admin/exports:
    post:
      operationId: createLegacyOrderExport
      tags: [Transactions]
      security: [{adminSession: [], csrfHeader: []}]
      parameters: [{$ref: "#/components/parameters/RequiredIdempotencyKey"}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/TransactionExportRequest"}}}}
      responses: {"200": {description: Receipt-backed CSV, content: {text/csv: {schema: {type: string}}}}, "400": {$ref: "#/components/responses/Error"}, "401": {$ref: "#/components/responses/Error"}, "403": {$ref: "#/components/responses/Error"}, "409": {$ref: "#/components/responses/Error"}}
  /api/admin/wechat-pay/order-exports:
    post:
      operationId: createLegacyWechatOrderExport
      tags: [Transactions]
      security: [{adminSession: [], csrfHeader: []}]
      parameters: [{$ref: "#/components/parameters/RequiredIdempotencyKey"}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/TransactionExportRequest"}}}}
      responses: {"200": {description: PII-safe local order CSV, content: {text/csv: {schema: {type: string}}}}, "400": {$ref: "#/components/responses/Error"}, "401": {$ref: "#/components/responses/Error"}, "403": {$ref: "#/components/responses/Error"}, "409": {$ref: "#/components/responses/Error"}}
  /api/admin/wechat-pay/orders:
    get:
      operationId: listLegacyWechatTransactions
      tags: [Transactions]
      security: [{adminSession: []}]
      parameters: [{$ref: "#/components/parameters/Limit"}]
      responses: {"200": {description: WeChat Pay order page, content: {application/json: {schema: {$ref: "#/components/schemas/TransactionOrderPage"}}}}, "401": {$ref: "#/components/responses/Error"}, "403": {$ref: "#/components/responses/Error"}}
  /api/admin/alipay/transactions:
    get:
      operationId: listLegacyAlipayTransactions
      tags: [Transactions]
      security: [{adminSession: []}]
      parameters: [{$ref: "#/components/parameters/Limit"}]
      responses: {"200": {description: Imported Alipay read-only page, content: {application/json: {schema: {$ref: "#/components/schemas/TransactionOrderPage"}}}}, "401": {$ref: "#/components/responses/Error"}, "403": {$ref: "#/components/responses/Error"}}
...
components:
  securitySchemes:
    adminSession: {type: apiKey, in: cookie, name: aicrm_admin_session}
    sidebarSession: {type: apiKey, in: cookie, name: aicrm_sidebar_session}
    csrfHeader: {type: apiKey, in: header, name: X-CSRF-Token}
    contextToken: {type: http, scheme: bearer}
    groupOpsWebhook: {type: apiKey, in: header, name: X-AICRM-Signature}
    paymentSession: {type: apiKey, in: cookie, name: aicrm_payment_session}
    wechatPaySignature: {type: apiKey, in: header, name: Wechatpay-Signature}
    operationCycleService: {type: http, scheme: bearer}
  parameters:
...
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 50}}
...
    RequiredIdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 8, maxLength: 200}}
...
  responses:
    Error: {description: Stable error response, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}}
    Unavailable: {description: Dependency or provider unavailable, content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}}
...
  schemas:
...
    TransactionOrder:
      type: object
      additionalProperties: true
      required: [id, record_origin, created_at, merchant_order_no, order_no, payer_name, payer_id, product_code, product_name, amount_yuan, currency, status, status_label, provider, provider_label, detail_url]
      properties:
        id: {$ref: "#/components/schemas/PositiveID"}
        record_origin: {type: string, enum: [native, v1_history]}
        created_at: {type: string, format: date-time}
        merchant_order_no: {type: string, maxLength: 200}
        order_no: {type: string, maxLength: 200}
        payer_name: {type: string, maxLength: 200}
        payer_id: {type: string, pattern: "^(customer:[1-9][0-9]*|)$"}
        product_code: {type: string, maxLength: 200}
        product_name: {type: string, maxLength: 500}
        amount_yuan: {type: string, pattern: "^[0-9]+[.][0-9]{2}$"}
        currency: {type: string, pattern: "^[A-Z]{3}$"}
        status: {type: string}
        status_label: {type: string}
        provider: {type: string, enum: [wechat, wechat_shop, alipay]}
        provider_label: {type: string}
        detail_url: {type: string, maxLength: 2048}
    TransactionOrderPage:
      type: object
      additionalProperties: false
      required: [items, orders, total, limit, has_more, next_cursor]
      properties:
        items: {type: array, items: {$ref: "#/components/schemas/TransactionOrder"}}
        orders: {type: array, items: {$ref: "#/components/schemas/TransactionOrder"}}
        total: {type: integer, minimum: 0}
        limit: {type: integer, minimum: 1, maximum: 100}
        has_more: {type: boolean}
        next_cursor: {type: string, maxLength: 512}
    TransactionExportRequest:
      type: object
      additionalProperties: false
      required: [resource, format]
      properties:
        resource: {type: string, const: orders}
        format: {type: string, const: csv}
        filter: {type: object, additionalProperties: true}
        filters: {type: object, additionalProperties: true}
...
    PositiveID: {type: integer, format: int64, minimum: 1}
...
    ErrorResponse:
      type: object
      properties: {ok: {type: boolean, const: false}, error: {type: string}, code: {type: string}, message: {type: string}, request_id: {type: string, minLength: 1}}
...